Do not replace custom link text
If a user added an _internal_ link with the [url] bbcode and custom text, do
not replace this custom text!

Example: the board has the domain `myforum.com`. A user creates a post
containing this bbcode:

```
[url=http://myforum.com/viewforum.php?f=1]my text[/url]
```

This is an internal link with an explicit custom link text "my text"; in this
case, the extension has to skip replacing this text.

closes #10